requirements.txt
yfinance==0.2.12
seaborn==0.11.2
numpy==1.21.5
plotly-express==0.4.1
stocknews==0.9.11